--- src/app/user/(settings)/reports/page.tsx
@@ -1,21 +1,0 @@
-"use client"
-import Modal from "../../../../components/modals/ModalReport"
-import React, { useState } from 'react';
-
-const Reports: React.FC = () => {
-  const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
-
-  const openModal = () => setIsModalOpen(true);
-  const closeModal = () => setIsModalOpen(false);
-
-  return (
-    <div>
-      <h1>Reportes</h1>
-      
-      <button onClick={openModal}>Abrir </button>
-      <Modal isOpen={isModalOpen} onClose={closeModal} />
-    </div>
-  );
-}
-
-export default Reports;
